+++ tapo-py/examples/tapo_p100.py
@@ -29,5 +29,6 @@ async def main():
     device_usage = await device.get_device_usage()
     print(f"Device usage: {device_usage.to_dict()}")
 
+
 if __name__ == "__main__":
     asyncio.run(main())