+++ tapo-py/src/handlers.rs
@@ -1,8 +1,10 @@
+mod color_light_handler;
 mod generic_device_handler;
 mod light_handler;
 mod plug_energy_monitoring_handler;
 mod plug_handler;
 
+pub use color_light_handler::*;
 pub use generic_device_handler::*;
 pub use light_handler::*;
 pub use plug_energy_monitoring_handler::*;