+++ tapo-py/src/lib.rs
@@ -5,13 +5,16 @@ mod handlers;
 use pyo3::prelude::*;
 
 use api_client::PyApiClient;
-use handlers::{PyEnergyDataInterval, PyPlugEnergyMonitoringHandler};
+use handlers::PyEnergyDataInterval;
+use tapo::requests::Color;
 
 #[pymodule]
 #[pyo3(name = "tapo")]
-fn tapo_py(_py: Python, m: &PyModule) -> PyResult<()> {
-    m.add_class::<PyApiClient>()?;
-    m.add_class::<PyPlugEnergyMonitoringHandler>()?;
-    m.add_class::<PyEnergyDataInterval>()?;
+fn tapo_py(_py: Python, module: &PyModule) -> PyResult<()> {
+    module.add_class::<PyApiClient>()?;
+
+    module.add_class::<PyEnergyDataInterval>()?;
+    module.add_class::<Color>()?;
+
     Ok(())
 }