+++ tapo/src/api/color_light_handler.rs
@@ -107,7 +107,7 @@ impl ColorLightHandler {
     ///
     /// # Arguments
     ///
-    /// * `color` - [crate::requests::Color]
+    /// * `color` - one of [crate::requests::Color] as defined in the Google Home app
     pub async fn set_color(&self, color: Color) -> Result<(), Error> {
         ColorLightSetDeviceInfoParams::new(&self.client)
             .color(color)