+++ tapo/src/requests/color.rs
@@ -4,8 +4,9 @@ use lazy_static::lazy_static;
 use serde::{Deserialize, Serialize};
 
 /// List of preset colors as defined in the Google Home app.
+#[derive(Debug, Clone, Eq, PartialEq, Hash, Serialize, Deserialize)]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
 #[allow(missing_docs)]
-#[derive(Debug, Eq, PartialEq, Hash, Serialize, Deserialize)]
 pub enum Color {
     CoolWhite,
     Daylight,